refactor types.go
